requirements.txt
psycopg2==2.9.6
py-cord==2.4.1
python-dotenv==1.0.0
rapidfuzz==3.1.1